pyproject.toml
[project]
name = "poker-counter"
version = "1.1.1"
description = "一个适用于中游（中国游戏）里的普通54张牌斗地主的记牌器。"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Natural Language :: Chinese (Simplified)",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.12",
    "Topic :: Games/Entertainment :: Board Games",
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "numpy>=2.2.4",
    "opencv-python-headless>=4.11.0.86",
    "pillow>=11.1.0",
    "pyyaml>=6.0.2",
]

[tool.pylint]
disable = ["W0201", "E0203", "C3001", "R0903", "E1101", "R0913", "R0917", "W0718", "C0103", "W0106", "W0212", "R0801"]

[tool.pyright]
pythonVersion = "3.12"
pythonPlatform = "Windows"
venvPath = "."
venv = ".venv"
extraPaths = ["src"]

[tool.ruff]
src = ["src"]
